# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "AccessTokenResponse": {
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "expires_in": {
            "title": "Expires In",
            "type": "integer"
          },
          "token_type": {
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/AuthenticatedUserResponse"
          }
        },
        "required": [
          "access_token",
          "expires_in",
          "user"
        ],
        "title": "AccessTokenResponse",
        "type": "object"
      },
...
      "ApiErrorDetail": {
        "properties": {
          "location": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "location",
          "message",
          "type"
        ],
        "title": "ApiErrorDetail",
        "type": "object"
      },
      "ApiErrorResponse": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "correlation_id": {
            "format": "uuid",
            "title": "Correlation Id",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ApiErrorDetail"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "details",
          "correlation_id"
        ],
        "title": "ApiErrorResponse",
        "type": "object"
      },
      "AuthenticatedContextResponse": {
        "properties": {
          "road_roles": {
            "items": {
              "$ref": "#/components/schemas/RoadRoleResponse"
            },
            "title": "Road Roles",
            "type": "array"
          },
          "user": {
            "$ref": "#/components/schemas/AuthenticatedUserResponse"
          }
        },
        "required": [
          "user",
          "road_roles"
        ],
        "title": "AuthenticatedContextResponse",
        "type": "object"
      },
      "AuthenticatedUserResponse": {
        "properties": {
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "display_name"
        ],
        "title": "AuthenticatedUserResponse",
        "type": "object"
      },
      "Body_login_for_access_token_v1_auth_token_post": {
        "properties": {
          "password": {
            "maxLength": 1024,
            "minLength": 1,
            "title": "Password",
            "type": "string"
          },
          "username": {
            "maxLength": 320,
            "minLength": 3,
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ],
        "title": "Body_login_for_access_token_v1_auth_token_post",
        "type": "object"
      },
...
      "RoadRoleResponse": {
        "properties": {
          "data_status": {
            "enum": [
              "real",
              "prepared",
              "simulated"
            ],
            "title": "Data Status",
            "type": "string"
          },
          "road_code": {
            "title": "Road Code",
            "type": "string"
          },
          "role": {
            "enum": [
              "manager",
              "supervisor"
            ],
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "road_code",
          "role",
          "data_status"
        ],
        "title": "RoadRoleResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/v1/auth/token"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "title": "ZENIT API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/auth/me": {
      "get": {
        "operationId": "read_authenticated_context_v1_auth_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticatedContextResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Request validation failed or the request cannot be processed.",
            "headers": {
              "X-Correlation-ID": {
                "description": "Request correlation identifier generated or accepted by the API.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "API error response.",
            "headers": {
              "X-Correlation-ID": {
                "description": "Request correlation identifier generated or accepted by the API.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Read Authenticated Context",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/token": {
      "post": {
        "operationId": "login_for_access_token_v1_auth_token_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_login_for_access_token_v1_auth_token_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessTokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Request validation failed or the request cannot be processed.",
            "headers": {
              "X-Correlation-ID": {
                "description": "Request correlation identifier generated or accepted by the API.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "The normalized login identifier is temporarily throttled.",
            "headers": {
              "Retry-After": {
                "description": "Seconds until another login attempt is allowed.",
                "schema": {
                  "minimum": 1.0,
                  "type": "integer"
                }
              },
              "X-Correlation-ID": {
                "description": "Request correlation identifier.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "API error response.",
            "headers": {
              "X-Correlation-ID": {
                "description": "Request correlation identifier generated or accepted by the API.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "Login For Access Token",
        "tags": [
          "auth"
        ]
      }
    },
...
  }
}